+++ test/python/pulse/test_schedule.py
@@ -23,10 +23,10 @@ from qiskit.pulse.channels import (DeviceSpecification, PulseChannelSpec, Qubit,
 from qiskit.pulse.commands import (FrameChange, Acquire, PersistentValue, Snapshot,
                                    functional_pulse, Instruction, AcquireInstruction,
                                    PulseInstruction, FrameChangeInstruction)
-from qiskit.pulse import pulse_lib
+from qiskit.pulse import pulse_lib, SamplePulse, CmdDef
 from qiskit.pulse.timeslots import TimeslotCollection, Interval
 from qiskit.pulse.exceptions import PulseError
-from qiskit.pulse.schedule import Schedule
+from qiskit.pulse.schedule import Schedule, ParameterizedSchedule
 from qiskit.test import QiskitTestCase
 
 
@@ -454,6 +454,45 @@ class TestSchedule(QiskitTestCase):
         sched = gp0(measure_chan) + Acquire(duration=10)(acquire_chan, memory_slot)
 
         self.assertEqual(sched.duration, 10)
+
+    def test_multiple_parameters_not_returned(self):
+        """Constructing ParameterizedSchedule object from multiple ParameterizedSchedules sharing
+        arguments should not produce repeated parameters in resulting ParameterizedSchedule
+        object."""
+        device = self.two_qubit_device
+
+        def my_test_par_sched_one(x, y, z):
+            result = PulseInstruction(
+                SamplePulse(np.array([x, y, z]), name='sample'),
+                device.drives[0]
+            )
+            return 0, result
+
+        def my_test_par_sched_two(x, y, z):
+            result = PulseInstruction(
+                SamplePulse(np.array([x, y, z]), name='sample'),
+                device.drives[0]
+            )
+            return 5, result
+
+        par_sched_in_0 = ParameterizedSchedule(
+            my_test_par_sched_one, parameters={'x': 0, 'y': 1, 'z': 2}
+        )
+        par_sched_in_1 = ParameterizedSchedule(
+            my_test_par_sched_two, parameters={'x': 0, 'y': 1, 'z': 2}
+        )
+        par_sched = ParameterizedSchedule(par_sched_in_0, par_sched_in_1)
+
+        cmd_def = CmdDef()
+        cmd_def.add('test', 0, par_sched)
+
+        actual = cmd_def.get('test', 0, 0.01, 0.02, 0.03)
+        expected = par_sched_in_0.bind_parameters(0.01, 0.02, 0.03) |\
+            par_sched_in_1.bind_parameters(0.01, 0.02, 0.03)
+        self.assertEqual(actual.start_time, expected.start_time)
+        self.assertEqual(actual.stop_time, expected.stop_time)
+
+        self.assertEqual(cmd_def.get_parameters('test', 0), ('x', 'y', 'z'))
 
 
 class TestScheduleWithDeviceSpecification(QiskitTestCase):