+++ CHANGELOG.md
@@ -96,6 +96,8 @@ The format is based on [Keep a Changelog].
 
 ### Fixed
 
+-   Fixed bug in `Pulse` for multiple parameters being added (\#2742)
+-   Fixed bug in `Pulse` for CmdDef arguments (\#2741)
 -   Fixed bug in `Operator` and `SuperOp` for initializing from circuit
     containing gates without an explicit matrix definition (\#2723)
 -   Possible to decompose SU(4) gate into non-CNOT basis with